realm_provider will need to return an instance.
Need to return a ready to use instance of the appropriate provider for the
realm, configured as per the settings for that realm, not just a package name.

As part of the multi-realm refactoring, all providers will be OO, and for each
realm, we'll have an instance of the provider class configured as per that
realm's settings.  (So, for instance, you could have two authentication realms,
both using the Database provider, but configured to talk to different
databases.)

Lots more work required before this can work.