Provider base class enhancements.
Add password matching here, so various providers can use $self->match_password()
and get crypted password handling for free.

Add stub methods for the methods which providers must implement, which will die
if they get called - so if a provider does not override a required method, it
will be clear what the issue is, rather than just exploding.

Handle the realm settings being passed to the constructor, stored in the object,
and add a `realm_settings` accessor.