New Config auth provider.
A simple auth provider which uses user accounts defined in the app's config file
for authentication.

This will be useful for testing multi-realm stuff easily, and will also
potentially be useful for people who want to add user authentication to their
app, but are happy to simply auth against user accounts listed in their app's
config for now.