Remove some method documentation.
Documentation on how to use the auth framework should be in the main module;
providers just need to implement the appropriate methods properly, but should
not need to document the methods or how to call them - just refer to the main
docs for that.

Providers should, IMO, focus on documenting how to set them up (configuration
example etc) and anything that is specific to that provider that the user should
know.