More tests - hashed passwords & redirect URL
Test that we can log in with a user whose password is stored as a hashed
password (as it should be in the real world, of course) and that it Just Works.

Also test that we can supply a return_url when we log in, and that we get
correctly rediretected there.