Probably shouldn't require Dancer::Plugin::Database.
The user might well be installing this to use the Config or Unix providers;
declaring D::P::Database as a pre-req will require them to have DBI too, which
is likely to be a pain if they're not planning to use a database otherwise.

I'm torn between including the most useful providers in the distribution, and
separating them out into individual dists, so this may change anyway.