Big refactor: stop using subroutine attributes.
This is a pretty big, backwards-incompatible change.

For the full rationale behind this change, see the discussion on the
Dancer-users mailing list:

http://lists.preshweb.co.uk/pipermail/dancer-users/2012-December/002808.html

Basically, multiple people whose opinions I respect have suggsted that using
subroutine attributes like this is a Bad Idea, and liable to being unreliable.

I've personally seen problems at work where attributes seem to disappear if you
try to run your code under Devel::Cover or the debugger, which is not very
helpful.

So, whilst this plugin is still young and not widely used, I'm planning to
change its interface.

Now, instead of saying:

    get '/secret' => sub :RequiresLogin { ... };

You'd say:

    get '/secret' => requires_login sub { ... };

And to require specific roles:

    get '/beer' => requires_role BeerDrinker => sub { ... };