Allow requires_role to take an arrayref of roles.
Let the user pass an arrayref of roles to requires_role, specifying that the
user should have one of those roles.

TODO: I'm not sure about this; perhaps an arrayref of roles passed to
requires_role should mean "must have all of these", and a new requires_any_role
should be added to handle the behaviour I'm adding here.

I wonder how common it is to require a user to have more than one role to access
something?  Is that "doing it wrong" when it comes to role-based access control?