Ensure username / password are not refs.
Handle the potential for JSON SQL injection.

As per the comment, if the username and password were deserialised from a JSON
request body, rather than form params, it's possible that they could be
hashrefs, instead of the straight scalars we'd expect.  And, if that hashref was
e.g. `{ '-like' => '%' }`, and was passed to a provider that uses a database
which in turn passed it to e.g. SQL::Abstract, we could end up with SQL rather
different to what we expected, matching more than we'd want.

So, check first that the username and password are straight scalars.