Cache logged in user details and roles.
Cache the details of the logged in user, and the roles they have, for the
duration of a request, so that repeated calls to `logged_in_user` or checking
roles don't cause a load of database queries.

@joshrabinowitz noticed this in particular with Dancer::Plugin::SimpleCRUD which
does various checks for roles, but it would apply to anything that calls
C<logged_in_user>, C<user_roles>, C<user_has_role> etc repeatedly.